requirements.txt
bitcoinlib 
ecdsa
python-bitcoinrpc